+++ src/electron/main.js
@@ -13,25 +13,24 @@ async function createWindow() {
             nodeIntegration: true,
             webSecurity: false,
             enableRemoteModule: false,
-            allowRunningInsecureContent: false
+            allowRunningInsecureContent: false,
+            contextIsolation: false
         },
         useContentSize: true
     })
 
-    const appUrl = process.env.ELECTRON_URL || url.format({
-        pathname: path.join(__dirname, 'app/index.html'),
-        protocol: 'file:',
-        slashes: true
-    })
+    const appUrl = process.env.ELECTRON_URL
+        ? path.join(process.env.ELECTRON_URL, '/webapp.html#/template')
+        : url.format({
+            pathname: path.join(__dirname, '/webapp.html#/template'),
+            protocol: 'file:',
+            slashes: true
+        })
 
     setupMenu(app, window)
     window.loadURL(appUrl)
     readVersion().then(version => console.log(`Running version ${version}`))
     setupIpcListeners(app, window)
-    // hot fix!
-    setTimeout(() => {
-        window.webContents.send('navigation', '/app/template')
-    }, 500)
 }
 
 app.whenReady().then(createWindow)