+++ src/ui/aboutPopup/AboutPopup.tsx
@@ -1,7 +1,5 @@
-import React, { Component } from 'react'
+import { Component } from 'react'
 import styles from './aboutPopup.module.scss'
-// eslint-disable-next-line
-import { bind } from 'helpful-decorators'
 import Popup from '../popup/Popup'
 import Authors from '../authors/Authors'
 import ExternalLink from '../misc/externalLink/ExternalLink'
@@ -23,7 +21,7 @@ class AboutPopup extends Component<{}, AboutPopupState> {
     }
 
     componentDidMount() {
-        ipcRenderer?.on('about-popup', (event, version: string) => {
+        ipcRenderer?.on('about-popup', (_event: any, version: string) => {
             this.setState({
                 active: true,
                 version: version
@@ -35,7 +33,6 @@ class AboutPopup extends Component<{}, AboutPopupState> {
         ipcRenderer?.removeListener('about-popup', () => {})
     }
 
-    @bind
     closeWindow() {
         this.setState({
             active: false
@@ -45,7 +42,7 @@ class AboutPopup extends Component<{}, AboutPopupState> {
     render () {
         if (this.state.active) {
             return (
-                <Popup title='About' closeHandler={this.closeWindow}>
+                <Popup title='About' closeHandler={() => this.closeWindow()}>
                     <div className={styles.container}>
                         <div className={styles.disclaimer}>
                             <p className={styles.paragraph}>Calligro lets you generate custom fonts from images created in graphics software like Gimp, Photoshop, Aseprite and others.</p>