+++ src/utils/unicodeRanges.ts
@@ -1,11 +1,7 @@
-export interface UnicodeRange {
-    category: string,
-    hexrange: [string, string]
-    range: [number, number]
-}
+import unicodeRanges, { UnicodeRange } from 'unicode-range-json'
+
 
 export function getUnicodeRanges(): UnicodeRange[] {
-    const unicodeRanges: UnicodeRange[] = require('unicode-range-json')
     unicodeRanges.splice(-2, 2)
     unicodeRanges.sort((a: UnicodeRange, b: UnicodeRange) => a.category.localeCompare(b.category, 'en'))